Fix routing, fix components
