+++ components/Containers/styles.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const StyledWrap = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+`
+
+export const StyledContainer = styled.div`
+  position: relative;
+  width: 620px;
+  justify-self: center;
+`
