+++ pages/settings/SettingsNavbar.tsx
@@ -1,38 +1,27 @@
-import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../store/store";
+import Link from "next/link";
+import { useRouter } from "next/router";
 import { SettingsNavbarItem } from "./SettingsNavbarItem";
 import { SettingsNavbarContainer } from "./styles";
-import { setSettingsNav } from "../../store/authorization/reducers";
 
 export const SettingsNavbar: React.FC = () => {
-  const authorization = useSelector((state: RootState) => state.authorizationSlice);
-  const dispatch = useDispatch();
-  
-  const navigateToPersonalInfo: React.MouseEventHandler<HTMLDivElement> = (e) => {
-    if (authorization.settingsNav !== 'personalInfo') {
-      dispatch(setSettingsNav('personalInfo'))
-    }
-  }
-  const navigateToChangePassword: React.MouseEventHandler<HTMLDivElement> = (e) => {
-    if (authorization.settingsNav !== 'changePassword') {
-      dispatch(setSettingsNav('changePassword'))
-    }
-  }
+  const router = useRouter();
 
   return (
     <SettingsNavbarContainer>
-      <SettingsNavbarItem
-        onClick={navigateToPersonalInfo}
-        active={authorization.settingsNav === 'personalInfo'}
-      >
-        Personal info
-      </SettingsNavbarItem>
-      <SettingsNavbarItem
-        onClick={navigateToChangePassword}
-        active={authorization.settingsNav === 'changePassword'}
-      >
-        Change password
-      </SettingsNavbarItem>
+      <Link href="/settings/update-data-form" passHref>
+        <a>
+          <SettingsNavbarItem active={router.route === '/settings/update-data-form'}>
+          Personal info
+          </SettingsNavbarItem>
+        </a>
+      </Link>
+      <Link href="/settings/update-password" passHref>
+        <a>
+          <SettingsNavbarItem active={router.route === '/settings/update-password'}>
+          Change password
+          </SettingsNavbarItem>
+        </a>
+      </Link>
     </SettingsNavbarContainer>
   )
 }
