+++ store/payment/actions.ts
@@ -1,10 +1,10 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import axios, { Subscribe } from '../../services/axios';
+import axios, { Checkout, Subscribe } from '../../services/axios';
 
 export const fetchBuy = createAsyncThunk(
   'authorizationSlice/fetchBuy',
-  async (priceId: number) => {
+  async ({ priceId, callBack }: Checkout ) => {
     const response = await axios.post<Subscribe>(`/payments/buy`, { priceId })
-    return response.data;
+    return { subscribe: response.data, callBack };
   }
 );
